+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) but modified for easier processing,
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Release 0.2.0 - 2026-05-14
+
+### Added
+
+ - **API Changes:** Support for recursive listing of files and directories in StorageServices.
+ - **API Changes:** Option to include directory entries in listing responses when supported by the server.
+ - Documentation for the library
+
 ## Release 0.1.0 - 2026-01-12
 
 ### Added